mysite/requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.7
async-timeout==4.0.2
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
configparser==5.3.0
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
django-admin-interface==0.19.2
django-appconf==1.0.5
django-colorfield==0.7.2
django-compressor==4.1
django-debug-toolbar==3.7.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-htmlmin==0.11.0
django-libsass==0.9
django-model-utils==4.2.0
django-notifications-hq==1.7.0
executing==0.9.1
html5lib==1.1
idna==3.4
ipython==8.4.0
jedi==0.18.1
jsonfield==3.1.0
kombu==5.2.4
libsass==0.21.0
matplotlib-inline==0.1.3
oauthlib==3.2.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.2.1
rcssmin==1.1.0
redis==4.3.5
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
swapper==1.3.0
Telethon==1.25.0
traitlets==5.3.0
tzdata==2022.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
